Insert into database

+++ src/object_model/foo.rs
@@ -1,11 +1,11 @@
 pub struct Foo {
-    pub id: i32,
+    pub id: i64,
     pub guid: String,
     pub name: String,
 }
 
 impl Foo {
-    pub fn new(id: i32, guid: &str, name: &str) -> Self {
+    pub fn new(id: i64, guid: &str, name: &str) -> Self {
         Self {
             id: id,
             guid: guid.into(),
